@@ -17,9 +17,8 @@
     <description>Payment Service for Tanzu Demo</description>
 
     <properties>
-        <java.version>11</java.version>
-        <spring-cloud.version>2021.0.4</spring-cloud.version>
-        <wavefront.version>2.3.0</wavefront.version>
+        <java.version>17</java.version>
+        <spring-cloud.version>2022.0.0-M5</spring-cloud.version>
     </properties>
 
     <dependencies>
