@@ -3,9 +3,11 @@ package com.vmware.tanzu.demo.payments;
 import io.micrometer.core.instrument.Counter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.cloud.stream.annotation.StreamListener;
+import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.function.Consumer;
 
 @Service
 public class PaymentConfirmationProcessor {
@@ -20,12 +22,13 @@ public class PaymentConfirmationProcessor {
         this.confirmationCounter = confirmationCounter;
     }
 
+    @Bean
     @Transactional
-    @StreamListener(PaymentChannels.CONFIRMATIONS)
-    public void confirmation(PaymentConfirmation paymentConfirmation) {
-        LOGGER.info("[Confirmation] - Payment [{}]", paymentConfirmation.getPaymentId());
-
-        this.paymentRepository.confirmPayment(paymentConfirmation.getPaymentId(), PaymentStatus.CONFIRMED);
-        this.confirmationCounter.increment();
+    public Consumer<PaymentConfirmation> confirmation() {
+        return confirmation -> {
+            LOGGER.info("[Confirmation] - Payment [{}]", confirmation.getPaymentId());
+            this.paymentRepository.confirmPayment(confirmation.getPaymentId(), PaymentStatus.CONFIRMED);
+            this.confirmationCounter.increment();
+        };
     }
 }
