@@ -1,22 +1,22 @@
 package com.vmware.tanzu.demo.payments;
 
+import jakarta.validation.constraints.NotNull;
+import org.springframework.cloud.stream.function.StreamBridge;
 import org.springframework.messaging.support.GenericMessage;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
-import javax.validation.constraints.NotNull;
-
 @Service
 public class PaymentMessaging {
 
-    private final PaymentChannels paymentChannels;
+    private final StreamBridge streamBridge;
 
-    public PaymentMessaging(PaymentChannels paymentChannels) {
-        this.paymentChannels = paymentChannels;
+    public PaymentMessaging(StreamBridge streamBridge) {
+        this.streamBridge = streamBridge;
     }
 
     @Async
     public void sendPaymentMessage(@NotNull final Payment payment) {
-        this.paymentChannels.payments().send(new GenericMessage<>(payment));
+        this.streamBridge.send("payments", new GenericMessage<>(payment));
     }
 }
